rename: study에서 studyMember(Attendance, GoalAchieve) 분리
