+++ src/main/java/com/grepp/spring/app/controller/api/StudyController.java
@@ -10,7 +10,7 @@ import com.grepp.spring.app.model.study.dto.request.StudyUpdateRequest;
 import com.grepp.spring.app.model.chat.service.ChatService;
 import com.grepp.spring.app.model.member.dto.response.ApplicantsResponse;
 import com.grepp.spring.app.model.member.dto.response.StudyMemberResponse;
-import com.grepp.spring.app.model.study.entity.Attendance;
+import com.grepp.spring.app.model.studyMmeber.entity.Attendance;
 import com.grepp.spring.app.model.member.service.MemberService;
 import com.grepp.spring.app.model.study.code.Category;
 import com.grepp.spring.app.model.study.code.Status;
@@ -19,10 +19,10 @@ import com.grepp.spring.app.model.study.dto.response.StudyInfoResponse;
 import com.grepp.spring.app.model.study.dto.response.StudyListResponse;
 import com.grepp.spring.app.model.study.dto.response.WeeklyAttendanceResponse;
 import com.grepp.spring.app.model.study.dto.response.WeeklyGoalStatusResponse;
-import com.grepp.spring.app.model.study.reponse.GoalsResponse;
-import com.grepp.spring.app.model.study.reponse.StudyNoticeResponse;
+import com.grepp.spring.app.model.study.dto.response.GoalsResponse;
+import com.grepp.spring.app.model.study.dto.response.StudyNoticeResponse;
 import com.grepp.spring.app.model.study.service.ApplicantService;
-import com.grepp.spring.app.model.study.service.StudyMemberService;
+import com.grepp.spring.app.model.studyMmeber.serivce.StudyMemberService;
 import com.grepp.spring.app.model.study.service.StudyService;
 import com.grepp.spring.infra.response.CommonResponse;
 import com.grepp.spring.infra.util.SecurityUtil;
