--- src/main/java/com/grepp/spring/app/model/member/code/StudyRole.java
@@ -1,5 +1,0 @@
-package com.grepp.spring.app.model.member.code;
-
-public enum StudyRole {
-    MEMBER, LEADER
-}
