+++ src/main/java/com/grepp/spring/app/model/study/repository/studyScheduleRepository/StudyScheduleRepository.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.model.study.repository.StudyScheduleRepository;
+package com.grepp.spring.app.model.study.repository.studyScheduleRepository;
 
 import com.grepp.spring.app.model.study.entity.StudySchedule;
 import org.springframework.data.jpa.repository.JpaRepository;
