+++ src/main/java/com/grepp/spring/app/model/study/service/StudyService.java
@@ -9,13 +9,13 @@ import com.grepp.spring.app.model.study.dto.response.WeeklyAchievementCount;
 import com.grepp.spring.app.model.alarm.code.AlarmType;
 import com.grepp.spring.app.model.alarm.service.AlarmService;
 import com.grepp.spring.app.model.auth.code.Role;
-import com.grepp.spring.app.model.member.code.StudyRole;
+import com.grepp.spring.app.model.studyMmeber.code.StudyRole;
 import com.grepp.spring.app.model.member.dto.response.ApplicantsResponse;
 import com.grepp.spring.app.model.member.dto.response.StudyMemberResponse;
 import com.grepp.spring.app.model.member.entity.Member;
-import com.grepp.spring.app.model.study.entity.StudyMember;
+import com.grepp.spring.app.model.studyMmeber.entity.StudyMember;
 import com.grepp.spring.app.model.member.repository.MemberRepository;
-import com.grepp.spring.app.model.member.repository.StudyMemberRepository;
+import com.grepp.spring.app.model.studyMmeber.repository.StudyMemberRepository;
 import com.grepp.spring.app.model.study.code.ApplicantState;
 import com.grepp.spring.app.model.study.code.DayOfWeek;
 import com.grepp.spring.app.model.study.code.GoalType;
@@ -26,14 +26,15 @@ import com.grepp.spring.app.model.study.dto.response.StudyInfoResponse;
 import com.grepp.spring.app.model.study.dto.response.StudyListResponse;
 import com.grepp.spring.app.model.study.dto.response.WeeklyGoalStatusResponse;
 import com.grepp.spring.app.model.study.entity.Applicant;
-import com.grepp.spring.app.model.study.entity.GoalAchievement;
+import com.grepp.spring.app.model.studyMmeber.entity.GoalAchievement;
 import com.grepp.spring.app.model.study.entity.Study;
 import com.grepp.spring.app.model.study.entity.StudyGoal;
-import com.grepp.spring.app.model.study.reponse.GoalsResponse;
-import com.grepp.spring.app.model.study.repository.ApplicantRepository;
-import com.grepp.spring.app.model.study.repository.GoalAchievementRepository;
-import com.grepp.spring.app.model.study.repository.StudyGoalRepository;
+import com.grepp.spring.app.model.study.dto.response.GoalsResponse;
+import com.grepp.spring.app.model.study.repository.applicant.ApplicantRepository;
+import com.grepp.spring.app.model.studyMmeber.repository.goalAchievement.GoalAchievementRepository;
+import com.grepp.spring.app.model.study.repository.goal.StudyGoalRepository;
 import com.grepp.spring.app.model.study.repository.StudyRepository;
+import com.grepp.spring.app.model.studyMmeber.serivce.StudyMemberService;
 import com.grepp.spring.infra.error.exceptions.AlreadyExistException;
 import com.grepp.spring.infra.error.exceptions.EarlierDateException;
 import com.grepp.spring.infra.error.exceptions.HasNotRightException;
