+++ src/main/java/com/grepp/spring/app/model/studyMmeber/repository/goalAchievement/GoalRepositoryCustomImpl.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.model.study.repository;
+package com.grepp.spring.app.model.studyMmeber.repository.goalAchievement;
 
 import com.grepp.spring.app.model.study.dto.response.CheckGoalResponse;
 import com.grepp.spring.app.model.study.dto.response.WeeklyAchievementCount;
@@ -6,7 +6,7 @@ import com.grepp.spring.app.model.member.entity.QStudyMember;
 import com.grepp.spring.app.model.study.entity.QGoalAchievement;
 import com.grepp.spring.app.model.study.entity.QStudyGoal;
 import com.grepp.spring.app.model.study.entity.StudyGoal;
-import com.grepp.spring.app.model.study.reponse.GoalsResponse;
+import com.grepp.spring.app.model.study.dto.response.GoalsResponse;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.core.types.dsl.NumberTemplate;
